# fix(xunix): also mount shared symlinked shared object files