--- a/Makefile
+++ b/Makefile
@@ -33,7 +33,6 @@ fmt/go:
 fmt/md:
 	go run github.com/Kunde21/markdownfmt/v3/cmd/markdownfmt@v3.1.0 -w ./README.md
 
-
 .PHONY: test
 test:
 	go test -v -count=1 ./...