--- a/xunix/gpu.go
+++ b/xunix/gpu.go
@@ -17,9 +17,10 @@ import (
 )
 
 var (
-	gpuMountRegex = regexp.MustCompile("(?i)(nvidia|vulkan|cuda)")
-	gpuExtraRegex = regexp.MustCompile("(?i)(libgl|nvidia|vulkan|cuda)")
-	gpuEnvRegex   = regexp.MustCompile("(?i)nvidia")
+	gpuMountRegex     = regexp.MustCompile("(?i)(nvidia|vulkan|cuda)")
+	gpuExtraRegex     = regexp.MustCompile("(?i)(libgl|nvidia|vulkan|cuda)")
+	gpuEnvRegex       = regexp.MustCompile("(?i)nvidia")
+	sharedObjectRegex = regexp.MustCompile(`\.so(\.[0-9\.]+)?$`)
 )
 
 func GPUEnvs(ctx context.Context) []string {
@@ -103,7 +104,7 @@ func usrLibGPUs(ctx context.Context, log slog.Logger, usrLibDir string) ([]mount
 				return nil
 			}
 
-			if filepath.Ext(path) != ".so" || !gpuExtraRegex.MatchString(path) {
+			if !sharedObjectRegex.MatchString(path) || !gpuExtraRegex.MatchString(path) {
 				return nil
 			}